requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.3
flask==2.3.2
flask-sqlalchemy==3.0.5
pillow==10.0.0
dlib==19.24.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
python-dateutil==2.8.2
cryptography==41.0.3
werkzeug==2.3.6
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.6
imutils==0.5.4
scipy==1.11.1
joblib==1.3.2
